+++ app/utils.py
@@ -15,50 +15,3 @@ def random_date(min_year=1900, max_year=datetime.now().year):
     end = start + timedelta(days=365 * years)
     return start + (end - start) * np.random.rand()
 
-
-class UnauthroizedException(HTTPException):
-    def __init__(self, detail: str, **kwargs):
-        super().__init__(status.HTTP_403_FORBIDDEN, detail=detail)
-
-
-class UnauthenticatedException(HTTPException):
-    def __init__(self):
-        super().__init__(
-            status_code=status.HTTP_401_UNAUTHORIZED, detail="Requires authentication"
-        )
-
-
-class VerifyToken:
-    def __init__(self):
-        jsonurl = f"https://{settings.auth0_domain}/.well-known/jwks.json"
-        self.jwks_client = jwt.PyJWKClient(jsonurl)
-
-    async def verify(
-        self,
-        security_scopes: SecurityScopes,
-        token: Optional[HTTPAuthorizationCredentials] = Depends(HTTPBearer()),
-    ):
-        if token is None:
-            raise UnauthenticatedException
-
-        try:
-            signing_key = self.jwks_client.get_signing_key_from_jwt(
-                token.credentials
-            ).key
-        except jwt.exceptions.PyJWKClientError as error:
-            raise UnauthroizedException(str(error))
-        except jwt.exceptions.DecodeError as error:
-            raise UnauthroizedException(str(error))
-
-        try:
-            payload = jwt.decode(
-                token.credentials,
-                signing_key,
-                algorithms=settings.auth0_algorithms,
-                audience=settings.auth0_api_audience,
-                issuer=settings.auth0_issuer,
-            )
-        except Exception as error:
-            raise UnauthroizedException(str(error))
-
-        return payload
